requirements.txt
flask
tensorflow-cpu
gunicorn
pillow
numpy
random